requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
imageio==2.34.1
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
networkx==3.3
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
tifffile==2024.6.18
tqdm==4.66.4
tzdata==2024.1
